refactor description quality app integration and add tests

+++ app/api/infrastructure/graph/model/models_gen.go
@@ -40,7 +40,7 @@ type InternalRepository struct {
 }
 
 type QualityProjectDesc struct {
-	Quality float64 `json:"quality"`
+	Quality int `json:"quality"`
 }
 
 type RemoveAPITokenInput struct {
