+++ app/api/infrastructure/kgservice/service.go
@@ -75,7 +75,7 @@ func (kg *kgService) GetGraph(ctx context.Context, description string) (entity.K
 	return entity.KnowledgeGraph{Items: items, Topics: topics}, nil
 }
 
-// GetItem gets a entity.KnowledgeGraphItem for a given id
+// GetItem gets a entity.KnowledgeGraphItem for a given id.
 func (kg *kgService) GetItem(ctx context.Context, id string) (entity.KnowledgeGraphItem, error) {
 	req := kgpb.GetItemReq{Id: id}
 
@@ -108,8 +108,8 @@ func (kg *kgService) GetItem(ctx context.Context, id string) (entity.KnowledgeGr
 	return item, nil
 }
 
-// DescriptionQuality gets the quality score of a given description
-func (kg *kgService) DescriptionQuality(ctx context.Context, description string) (float64, error) {
+// DescriptionQuality gets the quality score of a given description.
+func (kg *kgService) DescriptionQuality(ctx context.Context, description string) (int, error) {
 	req := kgpb.DescriptionQualityReq{Description: description}
 
 	res, err := kg.client.GetDescriptionQuality(ctx, &req)
@@ -117,7 +117,7 @@ func (kg *kgService) DescriptionQuality(ctx context.Context, description string)
 		return 0, err
 	}
 
-	return float64(res.QualityScore), nil
+	return int(res.QualityScore), nil
 }
 
 func stringToPointer(s string) *string {
