+++ app/api/usecase/kg/interactor.go
@@ -8,7 +8,7 @@ import (
 	"github.com/konstellation-io/kdl-server/app/api/pkg/logging"
 )
 
-// Interactor implements the UseCase interface.
+// interactor implements the UseCase interface.
 type interactor struct {
 	logger    logging.Logger
 	kgService kgservice.KGService
@@ -32,7 +32,7 @@ func (i *interactor) GetItem(ctx context.Context, id string) (entity.KnowledgeGr
 }
 
 // DescriptionQuality gets description quality.
-func (i *interactor) DescriptionQuality(ctx context.Context, description string) (float64, error) {
+func (i *interactor) DescriptionQuality(ctx context.Context, description string) (int, error) {
 	i.logger.Infof("Getting quality of description  \"%s\"", description)
 	return i.kgService.DescriptionQuality(ctx, description)
 }
